[refactor] : 코드 리팩토링 및 주석 추가
